Updates Snapshot test images
