+++ MLMiniTests/UIComponents/UIkit/GenericErrorScreen/MLGenericFeedbackScreenComponentSnapshotTests.swift
@@ -6,22 +6,14 @@
 //  Copyright © 2023 ale. All rights reserved.
 //
 
-import Foundation
-import XCTest
-import SnapshotTesting
 @testable import MLMini
 
-class MLGenericFeedbackScreenComponentSnapshotTests: XCTestCase {
+class MLGenericFeedbackScreenComponentSnapshotTests: MLBaseSnapshotTests {
     var feedbackView: MLGenericFeedbackScreenComponent!
-    var viewController: UIViewController!
-    
-    var snapshotTestTargetDevices: [String : ViewImageConfig] = [ "iPhone 8" :.iPhone8,
-                                                                  "iPhone Xs Max" :.iPhoneXsMax,
-                                                                  "iPad Pro 11" :.iPadPro11 ]
     
     override func setUp() {
+        super.setUp()
         feedbackView = MLGenericFeedbackScreenComponent()
-        viewController = UIViewController()
         viewController.view = feedbackView
     }
     
@@ -33,15 +25,5 @@ class MLGenericFeedbackScreenComponentSnapshotTests: XCTestCase {
     func testFeedback_Error_Screen() {
         feedbackView.type = .miscError
         verifyFeedbackViewSnapshot(withName: "FeedbackErrorViewSnapshot")
-    }
-    
-    private func verifyFeedbackViewSnapshot(withName: String) {
-        let result = snapshotTestTargetDevices.map { device in
-            verifySnapshot(matching: viewController,
-                           as: .image(on: device.value),
-                           named: "\(withName)_\(device.key)",
-                           record: false)
-        }
-        result.forEach { XCTAssertNil($0) }
     }
 }
